// src/api/schemas/userListItemSchema.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Keyo API
 * API for Keyo Polls app
 * OpenAPI spec version: 1.0.0
 */
import type { UserListItemSchemaAvatar } from './userListItemSchemaAvatar';
import type { UserListItemSchemaIsActiveMember } from './userListItemSchemaIsActiveMember';
import type { UserListItemSchemaJoinedAt } from './userListItemSchemaJoinedAt';
import type { UserListItemSchemaRole } from './userListItemSchemaRole';

export interface UserListItemSchema {
  id: number;
  username: string;
  display_name: string;
  avatar?: UserListItemSchemaAvatar;
  total_aura: number;
  created_at: string;
  role?: UserListItemSchemaRole;
  joined_at?: UserListItemSchemaJoinedAt;
  is_active_member?: UserListItemSchemaIsActiveMember;
}